# src/swagger.yaml
openapi: 3.0.0
info:
  title: Liquid
  description: Seamless authentication and user management APIs for your projects based on TypeScript, MongoDB and Redis.
  version: 0.1
paths:
...
  /user/code:
    get:
      summary: Generates a code to initiate a password reset.
      parameters:
        - in: query
          name: email
          schema:
            type: string
          description: Email ID of the user.
      responses:
        "200":
          description: Returns OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
      tags:
        - user-api
...
  /user/me:
    get:
      security:
        - bearerAuth: []
      summary: Retrieves the user profile associated with the API token.
      responses:
        "200": # status code
          description: Returns the user details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/definitions/user"
                      editableFields:
                        type: array
                        items:
                          type: string
      tags:
        - user-api

    patch:
      security:
        - bearerAuth: []
      summary: Updates the specified fields of the user profile associated with the API token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                middleName:
                  type: string
                lastName:
                  type: string
                password:
                  type: string
                bio:
                  type: string
      responses:
        "200": # status code
          description: Returns the user details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      ok:
                        type: boolean
      tags:
        - user-api
...
  /user/unfollow:
    post:
      security:
        - bearerAuth: []
      summary: Unfollows the user with the given ID from the account associated with the API token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target:
                  type: string
                  pattern: ^[6][a-f0-9]{23}$
      responses:
        "200":
          description: Returns OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
      tags:
        - user-api
...
  /user/following:
    get:
      security:
        - bearerAuth: []
      summary: Returns an array of users that the user associated with the API token is following.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Records per page.
        - in: query
          name: offset
          schema:
            type: string
            format: date-time
          description: createdAt field of last record in the previous page.
      responses:
        "200":
          description: Returns list of users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      records:
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                            approved:
                              type: boolean
                            target:
                              $ref: "#/definitions/user"
      tags:
        - user-api
  /user/followers:
    get:
      security:
        - bearerAuth: []
      summary: Returns an array of users that are following the user associated with the API token.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Records per page.
        - in: query
          name: offset
          schema:
            type: string
            format: date-time
          description: createdAt field of last record in the previous page.
      responses:
        "200":
          description: Returns list of users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      records:
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                            approved:
                              type: boolean
                            source:
                              $ref: "#/definitions/user"
      tags:
        - user-api
...
definitions:
  user:
    type: object
    properties:
      _id:
        type: string
      firstName:
        type: string
      middleName:
        type: string
      lastName:
        type: string
      gender:
        type: string
      username:
        type: string
      role:
        type: string
      bio:
        type: string
      designation:
        type: string
      profilePictureUrl:
        type: string
      pronouns:
        type: string
      verified:
        type: boolean
      varifiedDate:
        type: string
        format: date-time
      customLink:
        type: string
      followingCount:
        type: integer
      followerCount:
        type: integer
      isPrivate:
        type: boolean
      isSubscribed:
        type: boolean
      subscriptionTier:
        type: string
      subscriptionExpiry:
        type: string
        format: date-time
      isBanned:
        type: boolean
      isRestricted:
        type: boolean
      email:
        type: string
        format: email
      phone:
        type: string
      isFollowing:
        type: boolean
      requested:
        type: boolean